refactor(impact report): pulls out violation blocks into own component
