+++ client/src/components/ViolationSummaryBlock/components/ViolationCard/index.ts
